requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
openai>=1.50.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
